MAINT: Refactored fracture mechanics scripts to CLI
