+++ matscipy/cli/fracture_mechanics/meson.build
@@ -3,11 +3,14 @@
 # Pure Python sources
 python_sources = [
     '__init__.py',
-    'plot.py'
+    'quasistatic_crack.py',
+    'setup_crack.py',
+    'sinclair_continuation.py',
+    'sinclair_crack.py'
 ]
 
 # Install pure Python
 python.install_sources(
     python_sources,
-    subdir: 'matscipy/tool'
+    subdir: 'matscipy/cli/fracture_mechanics'
 )
