+++ pyproject.toml
@@ -70,13 +70,18 @@ cli = [
 # Generic
 matscipy-rms = "matscipy.cli.diffusion.rms:main"
 
-# Glasses
+# Structure generation
 matscipy-quench = "matscipy.cli.glasses.quench:main"
 
 # Electrochemistry
 matscipy-continuous2discrete = "matscipy.cli.electrochemistry.continuous2discrete:main"
 matscipy-poisson-nernst-planck = "matscipy.cli.electrochemistry.poisson_nernst_planck_solver:main"
 matscipy-stericify = "matscipy.cli.electrochemistry.stericify:main"
+
+# Fracture mechanics
+matscipy-quasistatic-crack = "matscipy.cli.fracture_mechanics.quasistatic_crack:main"
+matscipy-sinclair-continuation = "matscipy.cli.fracture_mechanics.sinclair_continuation:main"
+matscipy-sinclair-crack = "matscipy.cli.fracture_mechanics.sinclair_crack:main"
 
 # Interatomic potentials
 matscipy-average-eam-potential = "matscipy.cli.calculators.average_eam_potential:main"
